refactor(markdown): extract preprocessLaTeX into its own module
Move the preprocessLaTeX transform (and its nested processTabularContent
helper) out of the lib/utils.ts grab-bag into a dedicated, dependency-free
lib/preprocess-latex.ts, and relocate its unit tests to
lib/__tests__/preprocess-latex.test.ts. utils.ts imports the function for its
single internal caller; markdown.tsx and the markdown test import from the new
module. Behavior-preserving move — no logic changes.
