+++ components/markdown.tsx
@@ -5,7 +5,8 @@ import rehypeKatex from 'rehype-katex';
 import ReactMarkdown from 'react-markdown';
 
 import 'katex/dist/katex.min.css';
-import { preprocessLaTeX, cn } from '@/lib/utils';
+import { cn } from '@/lib/utils';
+import { preprocessLaTeX } from '@/lib/preprocess-latex';
 import { components } from './markdown/markdown-components';
 
 type Props = {
